requirements.txt
discord.py
pynput